server/requirements.txt
faiss==1.7.2
fastapi==0.92.0
h5py==3.8.0
keras==2.11.0
lshashpy3==0.0.8
numpy==1.24.1
Pillow==9.4.0
scikit_learn==1.2.1
